split out FileProcessor tests
